initial commit for tutorials ignoring sections in url
